Added some more stuff

+++ app/src/main/java/ru/nway/myweather/model/weather/Datum.java
@@ -19,7 +19,7 @@ public class Datum implements Serializable
     private double precipIntensityError;
     @SerializedName("precipProbability")
     @Expose
-    private int precipProbability;
+    private double precipProbability;
     @SerializedName("precipType")
     @Expose
     private String precipType;
@@ -73,7 +73,7 @@ public class Datum implements Serializable
         this.precipIntensityError = precipIntensityError;
     }
 
-    public int getPrecipProbability() {
+    public double getPrecipProbability() {
         return precipProbability;
     }
